vapi_bilingual/services/hr-stt/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
faster-whisper==1.0.3
python-dotenv==1.0.1
numpy==1.26.4
soundfile==0.12.1